bi apis

+++ Controller/category.js
@@ -1,6 +1,7 @@
-const client=require("../db")
-const message=require("../Helpers/message")
-
+const client	   = require("../db");
+const message	   = require("../Helpers/message");
+const redis 	   = require('redis');
+const redisClient  = redis.createClient(6379, 'localhost');
 
 /*** Create Categories ***/
 module.exports.createCategories = (user, info) => {
@@ -13,13 +14,30 @@ module.exports.createCategories = (user, info) => {
 						resolve(message.SOMETHINGWRONG+"01");
 					}else{
 						if (res1.rows == '') {
-								const sql = `INSERT INTO bi_categories(category_name,icon,is_status,created_by) VALUES ('${info.category_name}','${info.icon}','${1}','${user.id}')`;
+								const sql = `INSERT INTO bi_categories(category_name,icon,is_status,created_by) VALUES ('${info.category_name}','${info.icon}','${1}','${user.id}')RETURNING cat_id`;
 								client.query(sql, (error, result) => {
 									if(error){
 										resolve(message.SOMETHINGWRONG);
 									}else{
 										if (result != '') {
-											resolve(message.CREATED);
+											const redata = {
+												cat_id 	 	  : result.rows[0].cat_id,
+												category_name : info.category_name,
+												icon 		  : info.icon,
+												is_status	  : 1,
+												created_by    : user.id
+											}
+											redisClient.hmset('bi_categories', info.category_name, JSON.stringify(redata), function (err, data) {
+											    if(err){
+											    	resolve(message.SOMETHINGWRONG);
+											    }else{
+											    	if(data == 'OK'){
+												    	resolve(message.CREATED);
+											    	}else{
+												    	resolve(message.SOMETHINGWRONG);
+											    	}
+											    }
+											})
 										}else{
 											resolve(message.NOTCREATED);
 										}
@@ -70,7 +88,15 @@ module.exports.Categories = (user) => {
 									message : 'list of categories',
 									data     : catArray
 								}
-								resolve(response)
+								redisClient.hgetall('bi_categories', function (err, data) {
+								    if(err){
+								    	resolve(message.SOMETHINGWRONG);
+								    }else{
+								    	
+										resolve(response)
+								    }
+								})	
+								// resolve(response)
 							}else{
 								resolve(message.EMPTY)			
 							}
@@ -97,7 +123,15 @@ module.exports.Categories = (user) => {
 									message : 'list of categories',
 									data     : catArray
 								}
-								resolve(response)
+								redisClient.hgetall('bi_categories', function (err, data) {
+								    if(err){
+								    	resolve(message.SOMETHINGWRONG);
+								    }else{
+								    	
+										resolve(response)
+								    }
+								})	
+								//resolve(response)
 							}else{
 								resolve(message.EMPTY)			
 							}
@@ -118,17 +152,49 @@ module.exports.updateCategories = (user, info) => {
 		try{
 			if (user.role > 2) {
 				resolve(message.PERMISSIONERROR);
-			}else{				
-				const sql  = `UPDATE bi_categories SET category_name = '${info.category_name}', icon = '${info.icon}'  WHERE cat_id = '${info.cat_id}'`;
-				client.query(sql, (error, result) =>{
-					if (error) {
+			}else{		
+				const match = `SELECT * FROM bi_categories WHERE category_name = '${info.category_name}'`;
+		    	client.query(match, (matcherr, matchress) => {
+		    		if(matcherr){
 						resolve(message.SOMETHINGWRONG);
-					}else{
-						if (result) {
-							resolve(message.UPDATEDSUCCESS);
-						}
-					}
-				})
+		    		}else{
+		    			if(matchress.rows != ''){
+							resolve(message.DATANOTFOUND)
+		    			}else{
+		    				const sql  = `UPDATE bi_categories SET category_name = '${info.category_name}', icon = '${info.icon}'  WHERE cat_id = '${info.cat_id}'RETURNING cat_id`;
+							client.query(sql, (error, result) =>{
+								if (error) {
+									resolve(message.SOMETHINGWRONG);
+								}else{
+									if (result) {
+										const resdata = {
+											cat_id 	 	  : result.rows[0].cat_id,
+											category_name : info.category_name,
+											icon 		  : info.icon,
+											is_status	  : 1,
+											created_by    : user.id
+										}
+										redisClient.hmset('bi_categories', info.category_name, JSON.stringify(resdata), function (err, data) {
+										    if(err){
+										    	resolve(message.SOMETHINGWRONG);
+										    }else{
+										    	if(data == 'OK'){
+											    	resolve(message.UPDATEDSUCCESS);
+										    	}else{
+											    	resolve(message.NOTUPDATED);
+										    	}
+										    }
+										})	
+									}
+								}
+							})
+
+		    			}
+		    		}
+		    	})			
+
+
+				
 			}
 		}catch(error){
 			resolve(error)
@@ -143,18 +209,43 @@ module.exports.deleteCategories = (user, info) => {
 			if (user.role > 2) {
 				resolve(message.PERMISSIONERROR);
 			}else{
-				const sql  = `DELETE FROM bi_categories WHERE cat_id = '${info.cat_id}'`;
-				client.query(sql, (error, result) =>{
-					if (error) {
+				const chksql  = `SELECT * FROM bi_categories WHERE cat_id = '${info.cat_id}'`;
+				client.query(chksql, (chkerror, chkresult) =>{
+					if (chkerror) {
 						resolve(message.SOMETHINGWRONG);
 					}else{
-						if(result) {
-							resolve(message.DELETEDSUCCESS);
-						}else{		
-							resolve(message.NOTDELETED);
+						if (chkresult.rows == '') {
+							resolve(message.DATANOTFOUND);
+						}else{
+							const sql  = `DELETE FROM bi_categories WHERE cat_id = '${info.cat_id}'`;
+							client.query(sql, (error, result) =>{
+								if (error) {
+									resolve(message.SOMETHINGWRONG);
+								}else{
+									if(result) {
+									   const categoryname = chkresult.rows[0].category_name.trim();
+			                           redisClient.hdel('bi_categories',categoryname,function(err,redisdata){
+											if(err){
+												resolve(message.SOMETHINGWRONG);
+											}else{
+												if(redisdata == 1){
+													resolve(message.DELETEDSUCCESS);
+												}else{
+													resolve(message.NORDELETED);
+												}
+											}
+										})
+									}else{		
+										resolve(message.NOTDELETED);
+									}
+								}
+							})	
 						}
 					}
-				})
+				})	
+
+
+				
 			}
 		}catch(error){
 			resolve(error)
