+++ Controller/subcategories.js
@@ -1,6 +1,7 @@
-const client=require("../db")
-const message=require("../Helpers/message")
-
+const client 	   = require("../db");
+const message 	   = require("../Helpers/message");
+const redis 	   = require('redis');
+const redisClient  = redis.createClient(6379, 'localhost');
 
 
 /*** Create SubCategories ***/
@@ -14,13 +15,31 @@ module.exports.createSubCategories = (user, info) => {
 						resolve(message.SOMETHINGWRONG);
 					}else{
 						if (res1.rows == '') {	
-							    const sql = `INSERT INTO bi_subcategories(subcategory_name,cat_id,icon,is_status,created_by) VALUES ('${info.subcategory_name}','${info.cat_id}','${info.icon}','${1}','${user.id}')`;
+							    const sql = `INSERT INTO bi_subcategories(subcategory_name,cat_id,icon,is_status,created_by) VALUES ('${info.subcategory_name}','${info.cat_id}','${info.icon}','${1}','${user.id}')RETURNING subcat_id`;
 								client.query(sql, (error, result) => {
 									if(error){
 										resolve(message.SOMETHINGWRONG);
 									}else{
 										if (result != '') {
-											resolve(message.CREATEDSUCCESS);
+											const redata = {
+												subcat_id 	 	 : result.rows[0].subcat_id,
+												subcategory_name : info.subcategory_name,
+												cat_id 		     : info.cat_id,
+												icon 		     : info.icon,
+												is_status	     : 1,
+												created_by       : user.id
+											}
+											redisClient.hmset('bi_subcategories', info.subcategory_name, JSON.stringify(redata), function (err, data) {
+											    if(err){
+											    	resolve(message.SOMETHINGWRONG);
+											    }else{
+											    	if(data == 'OK'){
+												    	resolve(message.CREATEDSUCCESS);
+											    	}else{
+												    	resolve(message.SOMETHINGWRONG);
+											    	}
+											    }
+											})
 										}else{
 											resolve(message.NOTCREATED);
 										}
@@ -72,7 +91,15 @@ module.exports.SubCategories = (user, catId) => {
 									message : 'list of subcategories',
 									data     : subcatArray
 								}
-								resolve(response)
+								redisClient.hgetall('bi_subcategories', function (err, data) {
+								    if(err){
+								    	resolve(message.SOMETHINGWRONG);
+								    }else{
+								    	
+										resolve(response)
+								    }
+								})
+								//resolve(response)
 							}else{
 								resolve(message.EMPTY)			
 							}
@@ -100,7 +127,15 @@ module.exports.SubCategories = (user, catId) => {
 									message : 'list of subcategories',
 									data     : subcatArray
 								}
-								resolve(response)
+								redisClient.hgetall('bi_subcategories', function (err, data) {
+								    if(err){
+								    	resolve(message.SOMETHINGWRONG);
+								    }else{
+								    	
+										resolve(response)
+								    }
+								})
+								//resolve(response)
 							}else{
 								resolve(message.EMPTY)			
 							}
@@ -122,13 +157,33 @@ module.exports.updateSubCategories = (user, info) => {
 			if (user.role_id > 2) {
 				resolve(message.PERMISSIONERROR);
 			}else{				
-				const sql  = `UPDATE bi_subcategories SET subcategory_name = '${info.subcategory_name}', cat_id = '${info.cat_id}', icon = '${info.icon}' WHERE subcat_id = '${info.subcat_id}'`;
+				const sql  = `UPDATE bi_subcategories SET subcategory_name = '${info.subcategory_name}', cat_id = '${info.cat_id}', icon = '${info.icon}' WHERE subcat_id = '${info.subcat_id}'RETURNING subcat_id`;
 				client.query(sql, (error, result) =>{
+					// console.log(result);
 					if (error) {
 						resolve(message.SOMETHINGWRONG);
 					}else{
 						if (result) {
-							resolve(message.UPDATEDSUCCESS);
+							const resdata = {
+								subcat_id 	  : result.rows[0].subcat_id,
+								category_name : info.subcategory_name,
+								subcat_id 	  : info.subcat_id,
+								icon 		  : info.icon,
+								is_status	  : 1,
+								created_by    : user.id
+							}
+							redisClient.hmset('bi_subcategories', info.subcategory_name, JSON.stringify(resdata), function (err, data) {
+							    if(err){
+							    	resolve(message.SOMETHINGWRONG);
+							    }else{
+							    	if(data == 'OK'){
+								    	resolve(message.UPDATEDSUCCESS);
+							    	}else{
+								    	resolve(message.NOTUPDATED);
+							    	}
+							    }
+							})
+							//resolve(message.UPDATEDSUCCESS);
 						}
 					}
 				})
@@ -147,18 +202,41 @@ module.exports.deleteSubCategories = (user, info) => {
 			if (user.role_id > 2) {
 				resolve(message.PERMISSIONERROR);
 			}else{
-				const sql  = `DELETE FROM bi_subcategories WHERE subcat_id = '${info.subcat_id}'`;
-				client.query(sql, (error, result) =>{
-					if (error) {
+				const chksql  = `SELECT * FROM bi_subcategories WHERE subcat_id = '${info.subcat_id}'`;
+				client.query(chksql, (chkerr, chkres) =>{
+					if (chkerr) {
 						resolve(message.SOMETHINGWRONG);
 					}else{
-						if(result) {
-							resolve(message.DELETEDSUCCESS);
-						}else{		
-							resolve(message.NORDELETED);
+						if (chkres.rows == '') {
+							resolve(message.DATANOTFOUND)
+						}else{
+							const sql  = `DELETE FROM bi_subcategories WHERE subcat_id = '${info.subcat_id}'`;
+							client.query(sql, (error, result) =>{
+								if (error) {
+									resolve(message.SOMETHINGWRONG);
+								}else{
+									if(result) {
+										const subcategoryname = chkres.rows[0].subcategory_name.trim();
+			                            redisClient.hdel('bi_subcategories',subcategoryname,function(err,redisdata){
+											if(err){
+												resolve(message.SOMETHINGWRONG);
+											}else{
+												if(redisdata == 1){
+													resolve(message.DELETEDSUCCESS);
+												}else{
+													resolve(message.NORDELETED);
+												}
+											}
+										})
+										//resolve(message.DELETEDSUCCESS);
+									}else{		
+										resolve(message.NORDELETED);
+									}
+								}
+							})
 						}
 					}
-				})
+				})	
 			}
 		}catch(error){
 			resolve(error)
